feat: load more for category facet
